requirements.txt
absolufy-imports==0.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
Authlib==1.6.1
beautifulsoup4==4.13.4
bitarray==3.7.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.3
click==8.2.1
cloudpickle==3.1.1
cryptography==45.0.6
curl_cffi==0.13.0
cytoolz==1.0.1
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.116.1
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-adk==1.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.107.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.29.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
litellm==1.75.0
lxml==6.0.0
MarkupSafe==3.0.2
mcp==1.12.3
multidict==6.6.3
multitasking==0.0.12
narwhals==2.0.1
numpy==2.3.2
openai==1.99.1
opentelemetry-api==1.36.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
parsimonious==0.10.0
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-solc-x==2.0.4
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rlp==4.1.0
rpds-py==0.27.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
streamlit==1.48.0
tenacity==9.1.2
tiktoken==0.10.0
tokenizers==0.21.4
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
